refactor: split story-editor into focused modules and hooks
